+++ Tangem/Modules/Send/UnstakingModel.swift
@@ -121,7 +121,7 @@ private extension UnstakingModel {
             .withWeakCaptureOf(self)
             .flatMap { model, transaction in
                 model.sendTransactionDispatcher
-                    .sendPublisher(transaction: .staking(StakeKitMapper().mapToStakeKitTransaction(transaction)))
+                    .sendPublisher(transaction: .staking(StakingMapper().mapToStakeKitTransaction(transaction)))
                     .handleEvents(receiveOutput: { [weak model] output in
                         model?.proceed(transaction: transaction, result: output)
                     })
@@ -220,7 +220,7 @@ extension UnstakingModel: SendSummaryInput, SendSummaryOutput {
                     return nil
                 }
 
-                let stakeKitTx = model.mapToStakeKitTransaction(tx)
+                let stakeKitTx = StakingMapper().mapToStakeKitTransaction(tx)
                 return SendTransactionType.staking(stakeKitTx)
             }
             .eraseToAnyPublisher()
